requirements.txt
#DICA: USE O UV PIP INSTALL

# Core e Orquestração
langchain
crewai
langchain-community

# Integrações (Removendo travas rígidas para evitar o conflito de SDK)
langchain-openai
openai
langchain-google-genai
langchain-openrouter

# Ferramentas e Auxiliares
google-search-results
yfinance
duckduckgo-search
python-dotenv
litellm
pydantic
gitpython
bandit
safety